refactor: updated based on PR comments
