+++ src/main/kotlin/org/wfanet/measurement/securecomputation/controlplane/v1alpha/WorkItemsService.kt
@@ -17,9 +17,39 @@
 package org.wfanet.measurement.securecomputation.controlplane.v1alpha
 
 import com.google.protobuf.Message
+import io.grpc.Status
+import org.wfanet.measurement.securecomputation.controlplane.v1alpha.WorkItemsGrpcKt.WorkItemsCoroutineImplBase
 
-interface WorkItemsService {
 
-  fun publishMessage(queueName: String, message: Message)
+abstract class WorkItemsService: WorkItemsCoroutineImplBase() {
+
+  abstract fun publishMessage(queueName: String, message: Message)
+
+  override suspend fun createWorkItem(
+    request: CreateWorkItemRequest
+  ): WorkItem {
+
+    val workItem = request.workItem
+    val topicId = workItem.queue
+
+    try {
+      publishMessage(topicId, workItem.workItemParams)
+    } catch (e: Exception) {
+      throw when {
+        e.message?.contains("Topic id: $topicId does not exist") == true -> {
+          Status.NOT_FOUND
+            .withDescription(e.message)
+            .asRuntimeException()
+        }
+
+        else -> {
+          Status.UNKNOWN
+            .withDescription("An unknown error occurred: ${e.message}")
+            .asRuntimeException()
+        }
+      }
+    }
+    return workItem
+  }
 
 }
