+++ src/test/kotlin/org/wfanet/measurement/securecomputation/controlplane/v1alpha/GooglePubSubWorkItemsServiceTest.kt
@@ -95,6 +95,7 @@ class GooglePubSubWorkItemsServiceTest {
       subscriber.startAsync().awaitRunning()
       val result = deferred.await()
       assertThat(result).isEqualTo("test-user-name")
+      googlePubSubClient.deleteSubscription(projectId, subscriptionId)
       googlePubSubClient.deleteTopic(projectId, topicId)
     }
   }
@@ -150,6 +151,7 @@ class GooglePubSubWorkItemsServiceTest {
 
       subscriber.stopAsync().awaitTerminated()
       assertThat(receivedMessages.size).isEqualTo(numMessages)
+      googlePubSubClient.deleteSubscription(projectId, subscriptionId)
       googlePubSubClient.deleteTopic(projectId, topicId)
     }
   }
@@ -221,8 +223,11 @@ class GooglePubSubWorkItemsServiceTest {
         assertThat(messages.size).isEqualTo(messagesPerTopic)
       }
 
-      topics.forEach { topic ->
-        googlePubSubClient.deleteTopic(projectId, topic)
+      subscriptions.forEach {subscriptionId ->
+        googlePubSubClient.deleteSubscription(projectId, subscriptionId)
+      }
+      topics.forEach { topicId ->
+        googlePubSubClient.deleteTopic(projectId, topicId)
       }
 
     }
